+++ pkg/util/ip_test.go
@@ -7,6 +7,8 @@ import (
 	"net"
 	"reflect"
 	"testing"
+
+	"github.com/vmware-tanzu/nsx-operator/pkg/apis/vpc/v1alpha1"
 )
 
 func Test_rangesAbstractRange(t *testing.T) {
@@ -501,4 +503,25 @@ func TestCalculateIPFromCIDRs_IPv6(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestIPv4SubnetSizeApplicable(t *testing.T) {
+	assert.True(t, SubnetFamilyUsesIPv4(""))                             // unset → IPv4-only
+	assert.True(t, SubnetFamilyUsesIPv4(v1alpha1.IPAddressTypeIPv4))     // explicit IPv4
+	assert.True(t, SubnetFamilyUsesIPv4(v1alpha1.IPAddressTypeIPv4IPv6)) // dual-stack
+	assert.False(t, SubnetFamilyUsesIPv4(v1alpha1.IPAddressTypeIPv6))    // IPv6-only
+}
+
+func TestIPv6PrefixLengthApplicable(t *testing.T) {
+	assert.False(t, SubnetFamilyUsesIPv6(""))                            // unset → IPv4-only
+	assert.False(t, SubnetFamilyUsesIPv6(v1alpha1.IPAddressTypeIPv4))    // IPv4-only
+	assert.True(t, SubnetFamilyUsesIPv6(v1alpha1.IPAddressTypeIPv4IPv6)) // dual-stack
+	assert.True(t, SubnetFamilyUsesIPv6(v1alpha1.IPAddressTypeIPv6))     // IPv6-only
+}
+
+func TestSubnetFamilyUsesIPv6(t *testing.T) {
+	assert.False(t, SubnetFamilyUsesIPv6(""))
+	assert.False(t, SubnetFamilyUsesIPv6(v1alpha1.IPAddressTypeIPv4))
+	assert.True(t, SubnetFamilyUsesIPv6(v1alpha1.IPAddressTypeIPv6))
+	assert.True(t, SubnetFamilyUsesIPv6(v1alpha1.IPAddressTypeIPv4IPv6))
 }
