+++ test/e2e/nsx_subnet_test.go
@@ -58,7 +58,7 @@ func verifySubnetSetCR(subnetSet string) bool {
 	}
 	// IPv6PrefixLength is only set on SubnetSets for IPv6 or dual-stack clusters.
 	// Skip this check when the SubnetSet IP address type is IPv4-only (or unset).
-	if util.IPv6PrefixLengthApplicable(subnetSetCR.Spec.IPAddressType) {
+	if util.SubnetFamilyUsesIPv6(subnetSetCR.Spec.IPAddressType) {
 		wantIPv6 := vpcNetworkConfig.Spec.DefaultIPv6PrefixLength
 		if subnetSetCR.Spec.IPv6PrefixLength != wantIPv6 {
 			log.Error(nil, "IPv6PrefixLength mismatch", "IPv6PrefixLength", subnetSetCR.Spec.IPv6PrefixLength, "expected", wantIPv6)
@@ -158,7 +158,7 @@ func TestSubnetPrecreated(t *testing.T) {
 			SubnetWithAssociatedResourceAnnotation(t)
 		})
 		// Comment it, since it requires a long time to run.
-		//t.Run("case=PrecreatedSharedSubnetPoll", PrecreatedSharedSubnetPoll)
+		// t.Run("case=PrecreatedSharedSubnetPoll", PrecreatedSharedSubnetPoll)
 	})
 }
